+++ src/main/java/de/uib/configed/gui/ServerConfiguration.java
@@ -29,7 +29,8 @@ public class ServerConfiguration extends JTabbedPane {
 			.getPersistenceController();
 
 	public ServerConfiguration() {
-		panelHostConfig = new PanelHostConfig(this::setHostConfigTab, false);
+		panelHostConfig = new PanelHostConfig(this::setHostConfigTab,
+				persistenceController.getUserRolesConfigDataService().hasServerFullPermissionPD());
 
 		setHostConfigTab();