+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 ## [4.3.6.x] - 2025-xx-xx
 ### Fixed
 - Correct size in driver upload
-- `NPE` when reloading all data after changing groupaccess for client tree
-- Issue: user-roles not enforced correctly
+- Issue: `ClientSelectionDialog` UI not properly updated, when clicking on "Edit search" in `SavedSearchesDialog`
 
 ### Changed
+- Open new instance of `ClientSelectionDialog`, when clicking on "Create new search" in `SavedSearchesDialog`
+- `NPE` when reloading all data after changing groupaccess for client tree
+- Issue: user-roles not enforced correctly
 - Init tab components at start
 
 ## [4.3.6.7] - 2025-03-06